refactor NSTemplateTierController
